pyproject.toml
[project]
name = "noframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=0.9.9",
    "openai>=2.28.0",
    "black>=26.3.1",
    "isort>=8.0.1",
    "requests>=2.32.5",
]
